[#1]-Create interface for TextAnalysisService
